// src/client/common/src/models/problem-details-model.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

export interface ProblemDetailsModel {
    type?: string;
    title?: string;
    status?: number;
    detail?: string;
    instance?: string;
    extensions: { [key: string]: Object; };
}