// src/client/common/src/models/validation-problem-details-model.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

import type { ProblemDetailsModel } from "./problem-details-model";

export interface ValidationProblemDetailsModel extends ProblemDetailsModel {
    errors?: { [key: string]: string[]; };
}